+++ src/batou/lib/tests/test_nagios.py
@@ -6,8 +6,7 @@ def test_server_template(root):
         "http",
         command="http",
         args="-H localhost -u /login.html",
-        depend_on=[("localhost", "Supervisor")],
-    )
+        depend_on=[("localhost", "Supervisor")])
     root.component |= service
     server = NagiosServer()
     root.component += server
@@ -19,7 +18,7 @@ def test_server_template(root):
 #
 define service {
     use         generic-service
-    host_name   host
+    host_name   localhost
     service_description http
     check_command http!-H localhost -u /login.html
     servicegroups direct
@@ -28,7 +27,7 @@ define service {
 define servicedependency {
     use generic-servicedependency
     host_name localhost
-    dependent_host_name host
+    dependent_host_name localhost
     service_description Supervisor
     dependent_service_description http
 }