+++ src/batou/tests/test_config.py
@@ -1,10 +1,12 @@
-from batou.component import Component
-from batou.environment import Environment
-from batou import MissingEnvironment
-import batou
 import os
 import os.path
+
+import batou
 import pytest
+from batou import MissingEnvironment
+from batou.component import Component
+from batou.environment import Environment
+from batou.host import Host
 
 
 def test_parse_nonexisting_environment_raises_error(tmpdir):
@@ -25,8 +27,7 @@ def env():
 def env_prod():
     environment = Environment(
         "production",
-        basedir=os.path.dirname(__file__) + "/fixture/basic_service",
-    )
+        basedir=os.path.dirname(__file__) + "/fixture/basic_service")
     environment.load()
     return environment
 
@@ -45,11 +46,11 @@ def test_production_environment_is_loaded(env_prod):
     assert env_prod.host_domain == "example.com"
 
     assert sorted(env_prod.hosts) == [
-        "host1.example.com",
-        "host2.example.com",
-        "host3.example.com",]
+        "host1",
+        "host2",
+        "host3",]
 
-    host1 = env_prod.hosts["host1.example.com"]
+    host1 = env_prod.hosts["host1"]
     assert host1.name == "host1"
     assert host1.fqdn == "host1.example.com"
     assert host1.service_user == "alice"
@@ -70,7 +71,7 @@ def test_dev_environment_is_loaded(env):
         x.name for x in env.root_components if x.host is localhost])
     assert root_components == set(["zeo", "zope"])
 
-    zeo = env.get_root("zeo", "localhost")
+    zeo = env.get_root("zeo", env.hosts["localhost"])
     zeo.prepare()
     assert isinstance(zeo.component, Component)
     assert zeo.name == "zeo"
@@ -80,24 +81,24 @@ def test_dev_environment_is_loaded(env):
 def test_component_has_features_set(env):
     env.configure()
     # The localhost ZEO doesn't specify features and thus gets all set
-    zeo1 = env.get_root("zeo", "localhost").component
+    zeo1 = env.get_root("zeo", env.hosts["localhost"]).component
     assert zeo1.features == ["test", "test2"]
 
     # ZEO on host2 has selected only 'test'
-    zeo2 = env.get_root("zeo", "host2").component
+    zeo2 = env.get_root("zeo", env.hosts["host2"]).component
     assert zeo2.features == ["test"]
 
 
 def test_load_environment_with_overrides(env):
     env.overrides["zeo"] = {"port": "9002"}
     env.configure()
-    zeo = env.get_root("zeo", "localhost").component
+    zeo = env.get_root("zeo", env.hosts["localhost"]).component
     assert zeo.port == 9002
 
 
 def test_config_exceptions_orderable(env):
     env.configure()
-    c = env.get_root("zeo", "localhost").component
+    c = env.get_root("zeo", env.hosts['localhost']).component
 
     import sys
 
@@ -123,6 +124,7 @@ def test_config_exceptions_orderable(env):
         batou.CycleErrorDetected(ValueError()),
         batou.NonConvergingWorkingSet([c]),
         batou.DeploymentError(),
+        batou.DuplicateHostMapping("host", "map1", "map2"),
         batou.RepositoryDifferentError("asdf", "bsdf"),
         batou.DuplicateHostError("localhost"),
         batou.InvalidIPAddressError("asdf"),]